Split bond into a separate file
